backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
defusedxml==0.8.0rc2
discord==2.3.2
discord.py==2.4.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
oauthlib==3.2.2
propcache==0.2.1
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
pytest==8.3.3
pytest-django==4.9.0
python-slugify==8.0.4
pytest-asyncio==0.24.0
channels[daphne]==4.2.0
python-slugify==8.0.4
sqlparse==0.5.3
urllib3==2.2.3
whitenoise==6.8.2
yarl==1.18.3
pydantic==2.10.5
responses==0.25.6
Pillow==11.1.0